feat(bot): instance-as-agent with getInstanceCard() (#342)
feat(bot): add getInstanceCard() for instance-as-agent (#299)

Centralize instance presentation in lib/instance.ts. The instance
presents as a single agent externally, aggregating public skills
and hiding internals. Refactors /agent and /.well-known/agent.json
routes to use the shared getInstanceCard().
